requirements.txt
accelerate>=0.34.0
# bitsandbytes is required for QLoRA GPU training on Linux (HPC),
# but local macOS environments often do not have compatible wheels.
bitsandbytes>=0.43.1; platform_system == "Linux"
datasets>=2.20.0
evaluate>=0.4.2
huggingface_hub[cli]>=0.24.0
numpy>=1.26.0
openai>=1.40.0
pandas>=2.2.0
peft>=0.12.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
rouge-score>=0.1.2
scikit-learn>=1.5.0
scipy>=1.13.0
sentence-transformers>=3.0.1
torch>=2.2.0
tqdm>=4.66.0
transformers>=4.44.0
trl>=0.10.1
